+++ config.js
@@ -3,15 +3,19 @@ module.exports = {
     // http://conqueringthecommandline.com/book/ack_ag
     'ack',
     'ag',
-    // cmake is required to compile vim bundle YouCompleteMe
-    'cmake',
-    // Install GNU core utilities (those that come with OS X are outdated)
+    // https://github.com/wting/autojump
+    'autojump',
+    // alternative to `cat`: https://github.com/sharkdp/bat
+    'bat',
+    // Install GNU core utilities (those that come with macOS are outdated)
     // Don’t forget to add `$(brew --prefix coreutils)/libexec/gnubin` to `$PATH`.
     'coreutils',
     'dos2unix',
     // Install GNU `find`, `locate`, `updatedb`, and `xargs`, `g`-prefixed
     'findutils',
-    'fortune',
+    // 'fortune',
+    'fzf',
+    'readline', // ensure gawk gets good readline
     'gawk',
     // http://www.lcdf.org/gifsicle/ (because I'm a gif junky)
     'gifsicle',
@@ -19,12 +23,12 @@ module.exports = {
     // Install GNU `sed`, overwriting the built-in `sed`
     // so we can do "sed -i 's/foo/bar/' file" instead of "sed -i '' 's/foo/bar/' file"
     'gnu-sed --with-default-names',
+    // upgrade grep so we can get things like inverted match (-v)
+    'grep --with-default-names',
     // better, more recent grep
     'homebrew/dupes/grep',
     // https://github.com/jkbrzt/httpie
     'httpie',
-    'imagemagick',
-    'imagesnap',
     // jq is a sort of JSON grep
     'jq',
     // Mac App Store CLI: https://github.com/mas-cli/mas
@@ -38,10 +42,11 @@ module.exports = {
     // better/more recent version of screen
     'homebrew/dupes/screen',
     'tmux',
+    'todo-txt',
     'tree',
     'ttyrec',
     // better, more recent vim
-    'vim --with-override-system-vi',
+    'vim --with-client-server --with-override-system-vi',
     'watch',
     // Install wget with IRI support
     'wget --enable-iri'
@@ -55,17 +60,22 @@ module.exports = {
     //'comicbooklover',
     //'diffmerge',
     'dropbox',
+    'docker',
     'evernote',
     'flickr-uploadr',
     'gpgtools',
     'google-drive',
+    'flux',
+    'gpg-suite',
     //'ireadfast',
     'imazing',
     'iterm2',
     'little-snitch',
+    'macbreakz',
     'micro-snitch',
     'ngrok',
     'postico',
+    'signal',
     //'macvim',
     'sizeup',
     //'sketchup',
@@ -77,22 +87,22 @@ module.exports = {
     //'transmission',
     'transmit',
     'vlc',
+    'visual-studio-code',
     'xquartz'
   ],
   gem: [
-    'git-up'
   ],
   npm: [
     'antic',
     'buzzphrase',
     'eslint',
     'instant-markdown-d',
-    'npm-check',
-    'yo',
-    'generator-dockerize',
-    'gulp',
+    // 'generator-dockerize',
+    // 'gulp',
+    'npm-check-updates',
     'prettyjson',
     'trash',
     'vtop'
+    // ,'yo'
   ]
 };